+++ src/quicktions.pyx
@@ -23,7 +23,7 @@ This is an almost-drop-in replacement for the standard library's
 
 __all__ = ['Fraction']
 
-__version__ = '1.19'
+__version__ = '1.20'
 
 cimport cython
 from cpython.unicode cimport Py_UNICODE_TODECIMAL